oscore_message: avoid triggering assertions or UB
Even though successfully unprotected, data inside the message is still
treated as user input. Errors in the message must not cause a library
crash (via assert), but keep the program running with a suitable error
indication. As no error indication can be passed back here (but the
errors are expected to be caught earlier by the application, see
comment), the next best thing to do is to still initialize the data
behind the pointers with usable values.

References to NULL were replaced with actual pointer values (in theory
to the item after an array which is the last valid pointer value, in
practice to the first byte of the tag), as later users may pass those
into memcpy or similar that do not take NULL pointers lightly.