Remove payload_is_real workaround
To fully use this in combination with both states of MEMCPY mode, it'd
need to be a modifiable property, and given it's a struct and not
another indirection (which would require allocations in places I'd
rather not have them), it can not be used for communication between
subsequent operations on a oscore_msg_native_t.

Instead, payload_is_real=false mode is used all around.