Properly use payload_offset for mapping
This slightly decreases efficiency for mapping the payload the first
time in writable situations; the old code is left in there with an
explanation on why it's probably not much to be gained.

(The very old code that discovered the start manually is removed for
good, as that was already supposed to be superseeded by the memoized
offset).