Introduce flags in messages
The map_payload function, previously assumed to get away without that
distinction (but failing when a writable message does not (and why
should it) contain a payload marker), requires the distinction to be
made between writable and read-only (or "incomplete" and "parsable"?)
messages.

This does not introduce a type-based read/write distinction but only an
internal means of tracking which variables to look at.

Some other parts of the message's internals could be migrated to this
new mechanism, eg. payload_offset and class_e could be unioned with the
tag being the writable flag, and an explicit flag for "payload was
mapped, please memmove or reject adding options" added.