doc: Overhaul security context / message interaction description
The required identity of the unprotecte_request and prepare_response
security context was previously not stated.

The explanation was simplified by phrasing it as a case of the above
"sequence may require same value" rule, and only that that extends to
message modifications (but is in practic covered by the final function)
is special.