contexts: Split mutable and immutable parts
This allows parts of a security context to be used directly from
flash memory.

The changes to the plugtest-server example are illustrative in the
context construction, but incomplete in practice, as the prior usage
pattern of changing the user context details through the security
context pointer is rendered impossible by the const pointer. That
reflects the locking situation around the security contexts more
accurately, and is fixed in a subsequent commit that introduces actual
backing storage.

News-Section: API changes