plugtest demo: Make actual plug test keys optional
This allows running even in a RIOT setup that does have some algorithms
available, but not the ones needed for the plug test. The remaining demo
will still run, just the plugtest part itself is unavailable.

(This has become common now that AES-CCM is in liboscore but HKDF is not
yet).